pyproject.toml
[project]
name = "stockanalyser"
version = "0.1.0"
description = "AI-powered text analysis tool for financial insights"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "openai>=1.60.0",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
